Add files via upload
In this repose we solve many queries based on the selling of the pizza that which types of pizza of which category of pizza gives more revenue to the company. This is based on the data which we take and solve the queries with SQL. There are total 13 queries are in this project which are solved and make a diiferent file for each query.